reactor: componentize header
