+++ src/assets/components/styles/globals.scss
@@ -20,3 +20,4 @@ $screen-md: 768px;
 $screen-lg: 968px;
 $screen-xl: 1160px;
 $screen-grid: 1305.008px;
+$padding: 2.5rem;
