+++ src/components/organisms/Header/header.scss
@@ -20,42 +20,79 @@
     height: 2.5rem;
     position: relative;
     z-index: 1;
+
     @include for-size(screen-md) {
       height: 3rem;
     }
   }
 
-  &__menu {
-    // display: flex;
-    position: absolute;
-    left: -100%;
-    top: 0;
-    bottom: 0;
-    background-color: $black;
-    width: 100vw;
-    height: 100vh;
-    transition: 0.3s;
+  &__nav-hamburger {
+    display: flex;
+    justify-content: space-between;
 
-    &.active {
-      left: 0;
-      gap: 0.313rem;
+    & .header__menu {
+      position: absolute;
+      left: -100%;
+      top: 0;
+      bottom: 0;
+      padding-left: $padding;
+      background-color: $black;
+      width: 100vw;
+      height: 100vh;
+      transition: 0.3s;
+  
+      &.active {
+        left: 0;
+        gap: 0.313rem;
+      }
+      
+      @include for-size(screen-md) {
+        gap: 6.875rem;
+      }
+      
+      @include for-size(screen-md) {
+        position: static;
+        background-color: transparent;
+        width: auto;
+        height: 100%;
+      }
     }
-    
-    @include for-size(screen-md) {
-      gap: 6.875rem;
+
+    & .header__items {
+      display: flex;
+      flex-direction: column;
+      align-items: start;
+      margin-top: 20rem;
+  
+      @include for-size(screen-md) {
+        margin: 0.75rem;
+        flex-direction: row;
+      }
     }
-    
-    @include for-size(screen-md) {
-      position: static;
-      background-color: transparent;
-      width: auto;
-      height: 100%;
+
+    & .header__item {
+      // size: 2.5rem;
+      text-transform: uppercase;
+      // height: 100%;
+      // position: relative;
+      @include for-size(screen-md) {
+        font-size: 2rem;
+        text-transform: capitalize;
+      }
     }
   }
 
   &__nav {
     display: flex;
+    flex-wrap: wrap;
+    flex-direction: column;
+    gap: 2rem;
     justify-content: space-between;
+    align-items: center;
+
+    @include for-size(screen-sm) {
+      flex-direction: row;
+    }
   }
 
   &__hamburger {
@@ -94,10 +131,10 @@
   &__items {
     display: flex;
     flex-direction: column;
-    gap: 2rem;
-    margin-top: 20rem;
+    gap: 2.5rem;
+    align-items: center;
 
-    @include for-size(screen-md) {
+    @include for-size(screen-sm) {
       margin: .75rem;
       flex-direction: row;
     }
@@ -105,16 +142,15 @@
 
   &__item {
     font: {
-      size: 2.5rem;
+      size: 2rem;
       weight: bold;
     }
-    text-transform: uppercase;
-    height: 100%;
+    position: relative;
+
     @include for-size(screen-md) {
       font-size: 2rem;
       text-transform: capitalize;
     }
-    position: relative;
   
     &:after {
       content: "";
@@ -128,6 +164,7 @@
       transition: transform 0.3s ease-in-out;
       transform-origin: right;
     }
+
     &:hover:after {
       transform: translateX(-50%) scaleX(1.25);
       transform-origin: left;
